requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0
langchain-openai>=0.2.0
pydantic>=2.0
streamlit>=1.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
nest-asyncio>=1.6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
